Updates interface with get_factors, removes old plot  - Fixes interface with get_factors  - Removes S5, replaces with S6
